+++ app/main/views/notifications.py
@@ -39,11 +39,13 @@ from app.utils import (
     FAILURE_STATUSES,
     generate_notifications_csv,
     get_help_argument,
-    get_letter_printing_statement,
-    get_letter_validation_error,
     get_template,
     parse_filter_args,
     set_status_filters,
+)
+from app.utils.letters import (
+    get_letter_printing_statement,
+    get_letter_validation_error,
 )
 from app.utils.user import user_has_permissions
